refactor: extract Mobility tab to views/MobilityView.jsx + test (PR9)
Week-3 burndown (#52). Moves the `mobility` tab — the routines library
(expandable prime/foam-roll/yoga cards) and the consistency log — out of the
erg-dashboard.jsx monolith into a self-contained component.

- web/src/constants/logs.js: MOBILITY_ROUTINES (the routines library) and
  MOBILITY_STREAK_NOTE, lifted out of the monolith to sit with mobilityLog
  (the "library + tracking" pair the original comment describes), matching the
  week-2 precedent of co-locating a view's static data tables here.
- web/src/views/MobilityView.jsx: the mobility JSX, verbatim. Takes the lifted
  expand-state (mobOpen/setMobOpen) as props; imports its data from
  constants/logs.js.
- web/src/erg-dashboard.jsx: the ~315-line block becomes
  `{view === 'mobility' && <MobilityView ... />}`; the now-unused mobilityLog
  import is pruned.

Adds web/src/views/__tests__/MobilityView.test.jsx (5 tests): every routine
renders, clicking a card calls setMobOpen (open + collapse-to-null), the open
card reveals its block detail, and the streak note renders. Coverage: 100%
lines / 88.9% branch / 100% funcs. Also drops an unused `within` import from
the StrengthView test (lint warning).

Build, full Vitest suite (300 passing), lint, and format:check all green.

Closes #76
